Refactor cache upload functionality and improve test cases
