+++ packages/cache/__tests__/saveCacheV2.test.ts
@@ -1,14 +1,14 @@
 import * as core from '@actions/core'
 import * as path from 'path'
-import {saveCache} from '../src/cache'
+import { saveCache } from '../src/cache'
 import * as cacheUtils from '../src/internal/cacheUtils'
-import {CacheFilename, CompressionMethod} from '../src/internal/constants'
+import { CacheFilename, CompressionMethod } from '../src/internal/constants'
 import * as config from '../src/internal/config'
 import * as tar from '../src/internal/tar'
-import {CacheServiceClientJSON} from '../src/generated/results/api/v1/cache.twirp'
-import * as uploadCacheModule from '../src/internal/blob/upload-cache'
-import {BlobUploadCommonResponse} from '@azure/storage-blob'
-import {InvalidResponseError} from '../src/internal/shared/errors'
+import { CacheServiceClientJSON } from '../src/generated/results/api/v1/cache.twirp'
+import * as uploadCacheModule from '../src/internal/uploadUtils'
+import { BlobUploadCommonResponse } from '@azure/storage-blob'
+import { InvalidResponseError } from '../src/internal/shared/errors'
 
 let logDebugMock: jest.SpyInstance
 
@@ -28,11 +28,11 @@ jest.mock('@azure/storage-blob', () => ({
 
 beforeAll(() => {
   process.env['ACTIONS_RUNTIME_TOKEN'] = 'token'
-  jest.spyOn(console, 'log').mockImplementation(() => {})
-  jest.spyOn(core, 'debug').mockImplementation(() => {})
-  jest.spyOn(core, 'info').mockImplementation(() => {})
-  jest.spyOn(core, 'warning').mockImplementation(() => {})
-  jest.spyOn(core, 'error').mockImplementation(() => {})
+  jest.spyOn(console, 'log').mockImplementation(() => { })
+  jest.spyOn(core, 'debug').mockImplementation(() => { })
+  jest.spyOn(core, 'info').mockImplementation(() => { })
+  jest.spyOn(core, 'warning').mockImplementation(() => { })
+  jest.spyOn(core, 'error').mockImplementation(() => { })
   jest.spyOn(cacheUtils, 'resolvePaths').mockImplementation(async filePaths => {
     return filePaths.map(x => path.resolve(x))
   })
@@ -100,7 +100,7 @@ test('create cache entry failure', async () => {
 
   const createCacheEntryMock = jest
     .spyOn(CacheServiceClientJSON.prototype, 'CreateCacheEntry')
-    .mockReturnValue(Promise.resolve({ok: false, signedUploadUrl: ''}))
+    .mockReturnValue(Promise.resolve({ ok: false, signedUploadUrl: '' }))
 
   const createTarMock = jest.spyOn(tar, 'createTar')
   const finalizeCacheEntryMock = jest.spyOn(
@@ -116,8 +116,8 @@ test('create cache entry failure', async () => {
     .spyOn(cacheUtils, 'getArchiveFileSizeInBytes')
     .mockReturnValueOnce(archiveFileSize)
   const cacheVersion = cacheUtils.getCacheVersion(paths, compression)
-  const uploadCacheFileMock = jest
-    .spyOn(uploadCacheModule, 'uploadCacheFile')
+  const uploadCacheArchiveSDKMock = jest
+    .spyOn(uploadCacheModule, 'uploadCacheArchiveSDK')
     .mockReturnValueOnce(
       Promise.resolve({
         _response: {
@@ -139,7 +139,7 @@ test('create cache entry failure', async () => {
   expect(createTarMock).toHaveBeenCalledTimes(1)
   expect(getCompressionMock).toHaveBeenCalledTimes(1)
   expect(finalizeCacheEntryMock).toHaveBeenCalledTimes(0)
-  expect(uploadCacheFileMock).toHaveBeenCalledTimes(0)
+  expect(uploadCacheArchiveSDKMock).toHaveBeenCalledTimes(0)
 })
 
 test('finalize save cache failure', async () => {
@@ -152,12 +152,12 @@ test('finalize save cache failure', async () => {
   const createCacheEntryMock = jest
     .spyOn(CacheServiceClientJSON.prototype, 'CreateCacheEntry')
     .mockReturnValue(
-      Promise.resolve({ok: true, signedUploadUrl: signedUploadURL})
+      Promise.resolve({ ok: true, signedUploadUrl: signedUploadURL })
     )
 
   const createTarMock = jest.spyOn(tar, 'createTar')
 
-  const uploadCacheMock = jest.spyOn(uploadCacheModule, 'uploadCacheFile')
+  const uploadCacheMock = jest.spyOn(uploadCacheModule, 'uploadCacheArchiveSDK')
   uploadCacheMock.mockReturnValueOnce(
     Promise.resolve({
       _response: {
@@ -179,7 +179,7 @@ test('finalize save cache failure', async () => {
 
   const finalizeCacheEntryMock = jest
     .spyOn(CacheServiceClientJSON.prototype, 'FinalizeCacheEntryUpload')
-    .mockReturnValue(Promise.resolve({ok: false, entryId: ''}))
+    .mockReturnValue(Promise.resolve({ ok: false, entryId: '' }))
 
   const cacheId = await saveCache([paths], key)
 
@@ -218,7 +218,7 @@ test('save with uploadCache Server error will fail', async () => {
   jest
     .spyOn(CacheServiceClientJSON.prototype, 'CreateCacheEntry')
     .mockReturnValue(
-      Promise.resolve({ok: true, signedUploadUrl: signedUploadURL})
+      Promise.resolve({ ok: true, signedUploadUrl: signedUploadURL })
     )
 
   const archiveFileSize = 1024
@@ -226,7 +226,7 @@ test('save with uploadCache Server error will fail', async () => {
     .spyOn(cacheUtils, 'getArchiveFileSizeInBytes')
     .mockReturnValueOnce(archiveFileSize)
   jest
-    .spyOn(uploadCacheModule, 'uploadCacheFile')
+    .spyOn(uploadCacheModule, 'uploadCacheArchiveSDK')
     .mockRejectedValueOnce(new InvalidResponseError('boom'))
 
   const cacheId = await saveCache([paths], key)
@@ -241,14 +241,14 @@ test('uploadFile returns 500', async () => {
   jest
     .spyOn(CacheServiceClientJSON.prototype, 'CreateCacheEntry')
     .mockReturnValue(
-      Promise.resolve({ok: true, signedUploadUrl: signedUploadURL})
+      Promise.resolve({ ok: true, signedUploadUrl: signedUploadURL })
     )
 
   const archiveFileSize = 1024
   jest
     .spyOn(cacheUtils, 'getArchiveFileSizeInBytes')
     .mockReturnValueOnce(archiveFileSize)
-  jest.spyOn(uploadCacheModule, 'uploadCacheFile').mockRestore()
+  jest.spyOn(uploadCacheModule, 'uploadCacheArchiveSDK').mockRestore()
 
   uploadFileMock = jest.fn().mockResolvedValueOnce({
     _response: {
@@ -279,11 +279,11 @@ test('save with valid inputs uploads a cache', async () => {
   jest
     .spyOn(CacheServiceClientJSON.prototype, 'CreateCacheEntry')
     .mockReturnValue(
-      Promise.resolve({ok: true, signedUploadUrl: signedUploadURL})
+      Promise.resolve({ ok: true, signedUploadUrl: signedUploadURL })
     )
 
   const uploadCacheMock = jest
-    .spyOn(uploadCacheModule, 'uploadCacheFile')
+    .spyOn(uploadCacheModule, 'uploadCacheArchiveSDK')
     .mockReturnValueOnce(
       Promise.resolve({
         _response: {
@@ -300,7 +300,7 @@ test('save with valid inputs uploads a cache', async () => {
 
   const finalizeCacheEntryMock = jest
     .spyOn(CacheServiceClientJSON.prototype, 'FinalizeCacheEntryUpload')
-    .mockReturnValue(Promise.resolve({ok: true, entryId: cacheId.toString()}))
+    .mockReturnValue(Promise.resolve({ ok: true, entryId: cacheId.toString() }))
 
   const expectedCacheId = await saveCache([paths], key)
